face-api/requirements.txt
fastapi
uvicorn
efficientnet_pytorch
torchvision
torch
scikit-learn
numpy
pillow
python-multipart